refactor: Home/Board 페이지에서 중복으로 사용되는 RecruitCard 컴포넌트 분리
