+++ src/features/home/components/index.ts
@@ -1,5 +1,3 @@
-export { RecruitCard } from "@/features/home/components/RecruitCard";
-export type { RecruitCardProps } from "@/features/home/components/RecruitCard";
 export { RoommateProfileCard } from "@/features/home/components/RoommateProfileCard";
 export type { RoommateProfileCardProps } from "@/features/home/components/RoommateProfileCard";
 export { HomePageContent } from "@/features/home/components/HomePageContent";
